Use native structs to decode/encode
We used the serde serialize/deserialize from spdk before, but this means
that the struct format of a dependency affects whether or not we can
decode/encode our wallet. We should decouple this dependency by first
decoding to our own native struct type, and then converting into the
spdk struct type.