refactor: Make PinGuard a drop-in replacement of HomeScreen
Since HomeScreen is the only protected screen, we replace all
redirections to the home screen with a redirection to the PinGuard. The
PinGuard screen then decides whether to show the home screen, or a pin
setup screen.

We also refactor out the 'walletLoaded' variable from the PinGuard, as
this should not be handled by the pin guard widget.