Fix wallet not confirming spent outputs that we haven't spent ourselves
This is important in cases where users have their seed phrases imported
in 2 different wallets. If both wallets are at height N and wallet A
spends an output, then wallet B will not know this output is spent and
will not look into checking if it's confirmed.

This commits changes this behavior to also check unspent outputs for
confirmation.