Add new 'unknown spent' transaction history type
In some cases we might detect an output being spent despite the outgoing
transaction being unknown to us. To keep the transaction history from
showing wrong history info, we need to make it clear that a spending
transaction has taken place, even though we don't know the context for
it (recipient, txid, et cetera).

With this change, we make a single 'unknown spent' entry in the
transaction history when we detect outputs being spent. Since we don't
even know if these outputs are spent in the same transaction, we just
make a single entry for every block.